Got WRF test working, made CM1 Kmeans test a no-op while we figure out how to resolve an issue with the traces

+++ common/include/common/iris.h
@@ -35,7 +35,7 @@ typedef enum OPType {
 typedef struct DataDescriptor {
     CharStruct id_;
     size_t position_; // read/write start position
-    uint16_t size;
+    size_t size;
     uint chunk_index;
 };
